refactor: split settings mutation hooks
